+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This changelog keeps track of changes in a user-friendly way. It is based on [keep a changelog](https://keepachangelog.com/en/1.0.0/) by Olivier Lacan.
 
+## Unreleased
+### Added
+* Added full support for charged atoms in structures, along with default ionic radii for common ions.
+
+### Improved
+* The error message that occurs when an incomplete list of input arguments are provided in the CLI is now more descriptive.
+* Chemical formulas no longer contain the subscript "1" when an atom only occurs once.
+
+### Fixed
+* Fix an issue where the application would crash, when an invalid elements file was provided while attempting to load a structure file.
+
 ## [v1.0.0](https://github.com/molovol/MoloVol/releases/tag/v1.0.0) - 2021-09-14
 ### Added
 * Allow CIF files for structure input.